Refactor: remove unrelated functions from utils/helpers.js
