+++ utils/helpers.js
@@ -1,27 +1,6 @@
-const imgur = require('imgur-node-api')
-imgur.setClientID(process.env.IMGUR_CLIENT_ID)
-
 module.exports = {
-  shuffleArray: (array) => {
-    for (let i = array.length - 1; i > 0; i--) {
-      const j = Math.floor(Math.random() * (i + 1));
-      [array[i], array[j]] = [array[j], array[i]]
-    }
-  },
   isEmailValid: (email) => {
     const re = RegExp(/(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])/)
     return re.test(email)
-  },
-  uploadToImgur: (filePath) => {
-    return new Promise((resolve, reject) => {
-      imgur.upload(filePath, (err, img) => {
-        if (err) {
-          reject(err)
-        }
-        if (img) {
-          resolve(img.data.link)
-        }
-      })
-    })
   }
 }
